+++ create-meta/java/src/main/java/io/cucumber/createmeta/CreateMeta.java
@@ -21,7 +21,7 @@ import java.util.Map;
 import static io.cucumber.createmeta.VariableExpression.evaluate;
 import static java.nio.charset.StandardCharsets.UTF_8;
 
-public class CreateMeta {
+public final class CreateMeta {
     private static final JsonObject CI_DICT;
     private static final String JSON_PATH = "/io/cucumber/createmeta/ciDict.json";
 
@@ -31,6 +31,10 @@ public class CreateMeta {
         } catch (IOException e) {
             throw new RuntimeException("Unable to parse " + JSON_PATH, e);
         }
+    }
+
+    private CreateMeta(){
+
     }
 
     public static Meta createMeta(
@@ -73,6 +77,7 @@ public class CreateMeta {
         String url = evaluate(getString(ci, "url"), env);
         if (url == null) return null;
         JsonObject git = ci.get("git").asObject();
+        String buildNumber = evaluate(getString(ci, "buildNumber"), env);
         String remote = removeUserInfoFromUrl(evaluate(getString(git, "remote"), env));
         String revision = evaluate(getString(git, "revision"), env);
         String branch = evaluate(getString(git, "branch"), env);
@@ -81,12 +86,16 @@ public class CreateMeta {
         return new Ci(
                 name,
                 url,
+                buildNumber,
                 new Git(remote, revision, branch, tag)
         );
     }
 
     private static String getString(JsonObject json, String name) {
         JsonValue val = json.get(name);
+        if(val == null) {
+            throw new RuntimeException(String.format("Missing %s property in %s", name, json.toString()));
+        }
         return val.isNull() ? null : val.asString();
     }
 }