+++ gherkin-utils/javascript/test/messageTest.ts
@@ -4,14 +4,14 @@ import { Writable, pipeline } from 'stream'
 
 import { GherkinDocumentWalker } from '../src'
 import fs from 'fs'
-import glob from 'glob'
+import fg from 'fast-glob'
 import { promisify } from 'util'
 
 const asyncPipeline = promisify(pipeline)
 
 describe('Walking with messages', () => {
-  const localMessageFiles = glob.sync(`${__dirname}/messages/**/*.ndjson`)
-  const tckMessageFiles = glob.sync(
+  const localMessageFiles = fg.sync(`${__dirname}/messages/**/*.ndjson`)
+  const tckMessageFiles = fg.sync(
     `${__dirname}/../../../compatibility-kit/javascript/features/**/*.ndjson`
   )
   const messageFiles = [].concat(localMessageFiles, tckMessageFiles)