+++ gherkin/dart/lib/src/language/GherkinLanguageConstants.dart
@@ -1,0 +1,11 @@
+class GherkinLanguageConstants
+{
+  static const tagPrefix = '@';
+  static const commentPrefix = '#';
+  static const titleKeywordSeparator = ':';
+  static const tableCellSeparator = '|';
+  static const tableCellEscapeChar = '\\';
+  static const tableCellNewLineEscape = 'n';
+  static const docStringSeparator = '"""';
+  static const docStringAlternativeSeparator = '```';
+}