+++ react/javascript/src/components/gherkin/GherkinDocument.tsx
@@ -1,13 +1,11 @@
 import * as messages from '@cucumber/messages'
 import React from 'react'
-import Feature from './Feature'
+import { Feature } from './Feature'
 
 interface IProps {
   gherkinDocument: messages.GherkinDocument
 }
 
-const GherkinDocument: React.FunctionComponent<IProps> = ({ gherkinDocument }) => {
+export const GherkinDocument: React.FunctionComponent<IProps> = ({ gherkinDocument }) => {
   return gherkinDocument.feature ? <Feature feature={gherkinDocument.feature} /> : null
 }
-
-export default GherkinDocument