+++ react/javascript/src/components/gherkin/Rule.tsx
@@ -1,18 +1,15 @@
 import React from 'react'
-import Description from './Description'
-import Scenario from './Scenario'
-import * as messages from '@cucumber/messages'
-import Background from './Background'
-import Tags from './Tags'
-import Title from './Title'
-import Keyword from './Keyword'
-import HighLight from '../app/HighLight'
+import { Description } from './Description'
+import { Background } from './Background'
+import { Children } from './Children'
+import { Keyword } from './Keyword'
+import { Scenario } from './Scenario'
+import { Tags } from './Tags'
+import { Title } from './Title'
+import { HighLight } from '../app/HighLight'
+import { DefaultComponent, RuleProps, useCustomRendering } from '../customise'
 
-interface IProps {
-  rule: messages.Rule
-}
-
-const Rule: React.FunctionComponent<IProps> = ({ rule }) => {
+const DefaultRenderer: DefaultComponent<RuleProps, {}> = ({ rule }) => {
   return (
     <section>
       <Tags tags={rule.tags} />
@@ -21,7 +18,7 @@ const Rule: React.FunctionComponent<IProps> = ({ rule }) => {
         <HighLight text={rule.name} />
       </Title>
       <Description description={rule.description} />
-      <div className="cucumber-children">
+      <Children>
         {(rule.children || []).map((child, index) => {
           if (child.background) {
             return <Background key={index} background={child.background} />
@@ -31,9 +28,12 @@ const Rule: React.FunctionComponent<IProps> = ({ rule }) => {
             throw new Error('Expected background or scenario')
           }
         })}
-      </div>
+      </Children>
     </section>
   )
 }
 
-export default Rule
+export const Rule: React.FunctionComponent<RuleProps> = (props) => {
+  const ResolvedRenderer = useCustomRendering<RuleProps, {}>('Rule', {}, DefaultRenderer)
+  return <ResolvedRenderer {...props} />
+}