+++ suggest/javascript/src/index.ts
@@ -1,0 +1,4 @@
+export * from './types'
+export * from './buildStepDocuments'
+export * from './lspCompletionSnippet'
+export * from './index/index'