+++ suggest/javascript/src/index/index.ts
@@ -1,0 +1,4 @@
+export * from './types'
+export * from './bruteForceIndex'
+export * from './fuseIndex'
+export * from './jsSearchIndex'