+++ gherkin/CHANGELOG.md
@@ -11,15 +11,47 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 ### Added
 
+* [JavaScript] Expose `Errors` object.
+
 ### Changed
+
+* [Perl] Minimum Perl version upgraded to 5.12 (from 5.10.1)
+  ([#1665](https://github.com/cucumber/common/pull/1665) [ehuelsmann])
 
 ### Deprecated
 
 ### Removed
 
+* [Perl] Dependency on `IO::Scalar` removed
+  ([#1665](https://github.com/cucumber/common/pull/1665) [ehuelsmann])
+
+### Fixed
+
+## [20.0.1] - 2021-07-19
+
+### Changed
+
+* Update `messages` to v17.0.1
+
+## [20.0.0] - 2021-07-08
+
+### Changed
+
+* Update messages to v17.0.0
+* Update `rule` polish translation
+  ([#1579](https://github.com/cucumber/common/pull/1579) [l310][jakzal])
+* Add US Texan translations.
+  ([#1625](https://github.com/cucumber/common/pull/1625) [willmac321])
+* [Go] Move module paths to point to monorepo
+  ([#1550](https://github.com/cucumber/common/issues/1550))
+* [Ruby] Usage of Message DTOs instead of plain ruby hashes
+  ([#1603](https://github.com/cucumber/common/pull/1603))
+
 ### Fixed
 * [Ruby] Rules weren't inheriting the relevant tags during the Gherkin Query stage (Where it caches the NodeID)
   ([#1593](https://github.com/cucumber/cucumber/pull/1593) [luke-hill])
+
+* MDG parser correctly removes leading space from content in DocStrings
 
 ## [19.0.3] - 2021-05-24
 
@@ -57,7 +89,7 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 * [DotNet] Replaced Utf8Json with source-embedded TinyJson.
   ([#511](https://github.com/cucumber/cucumber/pull/1511)
-   [gasparnagy])   
+   [gasparnagy])
 
 * Update Dutch translation of the "Rule" keyword.
   ([#1513](https://github.com/cucumber/common/pull/1513)
@@ -88,7 +120,7 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 * [DotNet] Fixed .NET Gherkin compatibility with other Gherkin implementations.
   ([#511](https://github.com/cucumber/cucumber/pull/1511)
-   [gasparnagy])   
+   [gasparnagy])
 
 ## [18.1.1] - 2021-04-22
 
@@ -175,7 +207,7 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 ### Removed
 * [JavaScript] remove `IGherkinOptions#createReadStream`. This function was introduced in `9.1.0`
   in order to decouple this lib from the Node.js `fs` library. This decoupling was improved in
-  [#1333](https://github.com/cucumber/cucumber/pull/1333) which made it obsolete.  
+  [#1333](https://github.com/cucumber/cucumber/pull/1333) which made it obsolete.
   (Fixes [#1284](https://github.com/cucumber/cucumber/issues/1284)
 
 ### Fixed
@@ -1004,7 +1036,9 @@ to Gherkin 2.
 * First release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/gherkin/v19.0.3...main
+[Unreleased]: https://github.com/cucumber/cucumber/compare/gherkin/v20.0.1...main
+[20.0.1]:     https://github.com/cucumber/cucumber/compare/gherkin/v20.0.0...gherkin/v20.0.1
+[20.0.0]:     https://github.com/cucumber/cucumber/compare/gherkin/v19.0.3...gherkin/v20.0.0
 [19.0.3]:     https://github.com/cucumber/cucumber/compare/gherkin/v19.0.2...gherkin/v19.0.3
 [19.0.2]:     https://github.com/cucumber/cucumber/compare/gherkin/v19.0.1...gherkin/v19.0.2
 [19.0.1]:     https://github.com/cucumber/cucumber/compare/gherkin/v19.0.0...gherkin/v19.0.1
@@ -1080,12 +1114,14 @@ to Gherkin 2.
 [gasparnagy]:        https://github.com/gasparnagy
 [Haukinger]:         https://github.com/Haukinger
 [hkosova]:           https://github.com/hkosova
+[jakzal]:            https://github.com/jakzal
 [jargalan]:          https://github.com/jargalan
 [jmezach]:           https://github.com/jmezach
 [joscha]:            https://github.com/joscha
 [johnknoop]:         https://github.com/johnknoop
 [koterpillar]:       https://github.com/koterpillar
 [KniveX]:            https://github.com/KniveX
+[l310]:              https://github.com/l3l0/
 [l3pp4rd]:           https://github.com/l3pp4rd
 [LiohAu]:            https://github.com/LiohAu
 [luke-hill]:         https://github.com/luke-hill
@@ -1121,3 +1157,4 @@ to Gherkin 2.
 [zbmott]:            https://github.com/zbmott
 [Zearin]:            https://github.com/Zearin
 [WannesFransen1994]: https://github.com/WannesFransen1994
+[willmac321]:        https://github.com/willmac321